// src/establishment/dto/establishments-response.dto.ts
export interface EstablishmentResponse {
  id_estabelecimento: number;
  razao_social: string;
  nome_fantasia: string;
  cnpj: string;
  email: string;
  telefone: string;
  endereco: string;
  cidade: string;
  estado: string;
  agencia: string;
  conta: string;
  categoria: number;
  status: boolean;
  data_cadastro: string;
}